+++ pay-autotest/pay-autotest-threadpool/src/main/java/com/gigold/pay/autotest/threadpool/IfsysCheckThreadPool.java
@@ -38,7 +38,7 @@ public class IfsysCheckThreadPool {
 	private static final int CPUCORECOUNT = Runtime.getRuntime().availableProcessors();
 	private static final ExecutorService executor = Executors.newFixedThreadPool(CPUCORECOUNT + 1);
 
-	public TestResulteData execure() {
+	public TestResulteData execute() {
 		// 测试数据
 		TestResulteData testResulteData = new TestResulteData();
 		// 当前页
@@ -59,6 +59,28 @@ public class IfsysCheckThreadPool {
 		}
 
 		return testResulteData;
+	}
+	
+	
+	
+	public void initMock() {
+		// 当前页
+		int curPageNum = 1;
+		// 总页数
+		int pages = 1;
+		while (curPageNum <= pages) {
+			PageInfo<InterFaceInfo> pageInfo = interFaceService.getAllIfSys(curPageNum);
+			List<InterFaceInfo> ifsyslist = pageInfo.getList();
+			// 创建线程 用户初始化测试数据
+			Runnable worker = new InitMockTread(ifsyslist);
+			executor.execute(worker);
+			pages = pageInfo.getPages();
+			curPageNum++;
+		}
+		executor.shutdown();
+		while (!executor.isTerminated()) { //
+		}
+
 	}
 
 }
