+++ pay-autotest/pay-autotest-web/webapp/index.html
@@ -23,13 +23,13 @@
 		</div>
 		<form data-am-validator id="srearchFrom">
 			<div class="serchBox rightBtn">
-				系统：  <select data-am-selected id="ifSysId"></select>
-				产品： <select data-am-selected disabled id="ifProId"></select>
+				<select data-am-selected id="ifSysId"></select>
+				<select data-am-selected disabled id="ifProId"></select>
 				<input id="ifName" name="ifName" placeholder="输入需要查找的内容" />
 				<i class="am-icon-search" id="search"></i>
 			</div>
 		</form>
-		<div class="am-u-sm-8">
+		<div class="am-u-sm-12">
 			<div class="am-cf admin-main ">
 				<div class="admin-content ">
 					<table id="propDateShow"
@@ -46,7 +46,7 @@
 						</tbody>
 					</table>
 
-					<ul data-am-widget="pagination"
+					<ul data-am-widget="pagination" 
 						class="am-pagination am-pagination-default pageUl">
 					</ul>
 				</div>
